old_data/metadata_to_s3/requirements.txt
pyodbc
python-dotenv
pytest
pandas
pyarrow
boto3